+++ webpack.prod.js
@@ -1,0 +1,9 @@
+const webpack = require('webpack');
+const { merge } = require('webpack-merge');
+const common = require('./webpack.common.js');
+
+module.exports = merge(common, {
+  mode: 'production',
+  devtool: 'hidden-source-map',
+  plugins: [new webpack.EnvironmentPlugin({ NODE_ENV: 'production' })],
+});